% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{add_alpha}
\alias{add_alpha}
\title{Add alpha for a Rcolor}
\usage{
add_alpha(color, alpha = 0.3)
}
\arguments{
\item{color}{Rcolor}

\item{alpha}{alpha, default 0.3}
}
\description{
Add alpha for a Rcolor
}
