% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{add_analysis}
\alias{add_analysis}
\title{Add an analysis for a project}
\usage{
add_analysis(analysis_n, title = analysis_n, author = "Peng Chen", theme = 1)
}
\arguments{
\item{analysis_n}{analysis name}

\item{title}{file title}

\item{author}{author}

\item{theme}{1~10}
}
\value{
No return value
}
\description{
Add an analysis for a project
}
