% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{change_fac_lev}
\alias{change_fac_lev}
\title{Change factor levels}
\usage{
change_fac_lev(x, levels = NULL)
}
\arguments{
\item{x}{vector}

\item{levels}{custom levels}
}
\value{
factor
}
\description{
Change factor levels
}
\examples{
change_fac_lev(letters[1:5],levels = c("c","a"))
}
