% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{china_map}
\alias{china_map}
\title{Plot china map}
\usage{
china_map(china_shp = NULL, download_dir = "pcutils_temp", text_param = NULL)
}
\arguments{
\item{china_shp}{china.json file}

\item{download_dir}{download_dir, "pcutils_temp"}

\item{text_param}{parameters parse to \code{\link[ggplot2]{geom_text}}}
}
\value{
a ggplot
}
\description{
Plot china map
}
