% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{chunlian}
\alias{chunlian}
\title{Draw a Chunlian (Spring Festival couplet) using ggplot2}
\usage{
chunlian(
  words = NULL,
  bg_size = 20,
  bg_shape = 22,
  bg_fill = "red2",
  text_size = 10,
  text_params = list(),
  font_file = NULL,
  download_dir = "pcutils_temp"
)
}
\arguments{
\item{words}{A character vector containing three strings for the three lines of the couplet}

\item{bg_size}{Size of the points in geom_point, 20}

\item{bg_shape}{Shape of the points in geom_point (21~25), 22 or 23 are very good.}

\item{bg_fill}{Fill color of the points in geom_point}

\item{text_size}{Size of the text in geom_text, 10}

\item{text_params}{parameters parse to geom_text}

\item{font_file}{font file, e.g XX.ttf, XX.ttc}

\item{download_dir}{download_dir for font_file}
}
\value{
A ggplot object representing the Chunlian
}
\description{
Draw a Chunlian (Spring Festival couplet) using ggplot2
}
