% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count2}
\alias{count2}
\title{Like uniq -c in shell}
\usage{
count2(plot_df)
}
\arguments{
\item{plot_df}{two columns: first is type, second is number}
}
\description{
Like uniq -c in shell
}
\examples{
count2(data.frame(group= c("A", "A", "B", "C", "C", "A"),value =c(2, 2, 2, 1, 3, 1)))
}
