% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{dabiao}
\alias{dabiao}
\title{Print with =}
\usage{
dabiao(str = "", ..., n = 80, char = "=", mode = c("middle", "left", "right"))
}
\arguments{
\item{str}{output strings}

\item{...}{strings will be paste together}

\item{n}{the number of output length}

\item{char}{side chars default:=}

\item{mode}{"middle", "left" or "right"}
}
\description{
Print with =
}
