% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{del_ps}
\alias{del_ps}
\title{Detach packages}
\usage{
del_ps(p_list, ..., origin = NULL)
}
\arguments{
\item{p_list}{a vector of packages list}

\item{...}{packages}

\item{origin}{keep the original Namespace}
}
\value{
No return value
}
\description{
Detach packages
}
