% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{df2link}
\alias{df2link}
\title{df 2 link}
\usage{
df2link(test, fun = sum)
}
\arguments{
\item{test}{df}

\item{fun}{function to summary the elements number, defalut: \code{sum}, you can choose \code{mean}.}
}
\description{
df 2 link
}
\examples{
data(otutab)
cbind(taxonomy,num=rowSums(otutab))[1:10,]->test
df2link(test)

}
