% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df2net}
\alias{df2net}
\title{Transfer a dataframe to a network edgelist.}
\usage{
df2net(test)
}
\arguments{
\item{test}{df}
}
\value{
igraph
}
\description{
Transfer a dataframe to a network edgelist.
}
\examples{
cbind(taxonomy,num=rowSums(otutab))[1:20,]->test
df2net(test)->ttt
library(igraph)
plot(ttt,vertex.color=MetaNet:::tidai(V(ttt)$level,get_cols(7)),layout=layout_as_tree(ttt),vertex.size=mmscale(V(ttt)$weight,5,13))
}
