% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{distance2df}
\alias{distance2df}
\title{Convert a distance matrix to a data frame}
\usage{
distance2df(distance_matrix)
}
\arguments{
\item{distance_matrix}{A distance matrix where rows and columns are all unique names from 'from' and 'to' columns.}
}
\value{
A data frame containing three columns: from, to, count.
}
\description{
This function converts a distance matrix into a data frame with three columns: from, to, count.
The rows and columns of the matrix are all unique names from the 'from' and 'to' columns,
}
\examples{
distance_matrix <- matrix(c(0, 1, 2, 3, 4, 5, 6, 7, 8), nrow = 3)
distance2df(distance_matrix)
}
