% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{DNA_plot}
\alias{DNA_plot}
\title{Plot a DNA double helix}
\usage{
DNA_plot(
  col_DNA = "#377EB8",
  col_ATCG = c("#7FC97F", "#FB8072", "#FFFFB3", "#A6CEE3"),
  DNA_length = 2
)
}
\arguments{
\item{col_DNA}{col_DNA, "#377EB8"}

\item{col_ATCG}{col_ATCG, c("#7FC97F","#FB8072","#FFFFB3","#A6CEE3")}

\item{DNA_length}{DNA_length, 2}
}
\value{
ggplot
}
\description{
Plot a DNA double helix
}
\examples{
DNA_plot()
}
\references{
\code{https://github.com/SherryDong/create_plot_by_R_base}
}
