% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{download2}
\alias{download2}
\title{Download File}
\usage{
download2(url, file_path, timeout = 300, force = FALSE)
}
\arguments{
\item{url}{The URL from which to download the file.}

\item{file_path}{The full path to the file.}

\item{timeout}{timeout, 300s}

\item{force}{FALSE, if TRUE, overwrite existed file}
}
\description{
This function downloads a file from the provided URL and saves it to the specified location.
}
