% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{explode}
\alias{explode}
\title{Explode a dataframe if there are split charter in one column}
\usage{
explode(df, column, split = ",")
}
\arguments{
\item{df}{dataframe}

\item{column}{column}

\item{split}{split string}
}
\description{
Explode a dataframe if there are split charter in one column
}
\examples{
\dontrun{
#df <- data.frame(a = 1:2, b = c("a,b", "c"), c = 3:4)
#explode(df, "b", ",")
}
}
