% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_labels}
\alias{generate_labels}
\title{Generate labels position}
\usage{
generate_labels(
  labels = NULL,
  input = c(0, 0),
  nrows = NULL,
  ncols = NULL,
  x_offset = 0.3,
  y_offset = 0.15,
  just = 1
)
}
\arguments{
\item{labels}{labels}

\item{input}{c(0,0)}

\item{nrows}{default: NULL}

\item{ncols}{default: NULL}

\item{x_offset}{0.3}

\item{y_offset}{0.15}

\item{just}{0~5}
}
\value{
matrix
}
\description{
Generate labels position
}
\examples{
library(ggplot2)
labels <- vapply(1:8, \(i)paste0(sample(LETTERS, 4), collapse = ""), character(1))
df <- data.frame(label = labels, generate_labels(labels))
ggplot(data = df) +
    geom_label(aes(x = X1, y = X2, label = label))
}
