% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cols}
\alias{get_cols}
\title{Get n colors}
\usage{
get_cols(n, pal = NULL)
}
\arguments{
\item{n}{how many colors you need}

\item{pal}{a vector of colors, you can get from here too.{RColorBrewer::brewer.pal(5,"Set2")} {ggsci::pal_aaas()(5)}}
}
\value{
n colors
}
\description{
Get n colors
}
\examples{
get_cols(10)->my_cols
scales::show_col(my_cols)
}
