% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{get_doi}
\alias{get_doi}
\title{Download supplemental materials according to a doi}
\usage{
get_doi(
  doi,
  dir = "~/Downloads/",
  bget_path = "~/software/bget_0.3.2_Darwin_64-bit/bget"
)
}
\arguments{
\item{doi}{doi}

\item{dir}{dir}

\item{bget_path}{your bget_path}
}
\value{
file at work directory
}
\description{
Download supplemental materials according to a doi
}
