% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_legend2}
\alias{get_legend2}
\title{Get a legend from a ggplot object}
\usage{
get_legend2(plot, legend = NULL)
}
\arguments{
\item{plot}{a ggplot object}

\item{legend}{NULL, or position ("top")}
}
\value{
a grob object, or NULL if no legend found
}
\description{
Get a legend from a ggplot object
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg, color = mpg)) +
  geom_point()
legend <- get_legend2(p)
plot(legend)
}
