% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Heatmap by ggplot}
\usage{
ggheatmap(
  otutab,
  pal = NULL,
  scale = "none",
  rowname = TRUE,
  colname = TRUE,
  row_cluster = TRUE,
  col_cluster = TRUE,
  row_annotation = NULL,
  col_annotation = NULL,
  annotation_pal = NULL
)
}
\arguments{
\item{otutab}{otutab}

\item{pal}{the main color pal, a vector of colors}

\item{scale}{"none", "row", "column"}

\item{rowname}{show row names?}

\item{colname}{show column names?}

\item{row_cluster}{cluster the row?}

\item{col_cluster}{cluster the column?}

\item{row_annotation}{row annotation}

\item{col_annotation}{column annotation}

\item{annotation_pal}{the annotation color pal, a list. e.g. list(Group=c("red","blue"))}
}
\value{
a ggplot
}
\description{
Heatmap by ggplot
}
\examples{
data(otutab)
ggheatmap(otutab[1:30, ],
    scale = "row", row_annotation = otutab[1:30, 1:2],
    col_annotation = metadata[, c(2, 4)]
)
}
