% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gghist}
\alias{gghist}
\title{gg Histogram}
\usage{
gghist(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{parameters parse to \code{\link[ggpubr]{gghistogram}}}
}
\value{
ggplot
}
\description{
gg Histogram
}
\examples{
gghist(rnorm(100))
}
