% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{give_you_a_rose}
\alias{give_you_a_rose}
\title{Give you a rose}
\usage{
give_you_a_rose(color = "red3")
}
\arguments{
\item{color}{"skyblue3"}
}
\value{
plot
}
\description{
Give you a rose
}
\references{
https://mp.weixin.qq.com/s/W-BYPR3UXL120XWpTmN3rA
}
