% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{grepl.data.frame}
\alias{grepl.data.frame}
\title{Grepl applied on a data.frame}
\usage{
grepl.data.frame(pattern, x, ...)
}
\arguments{
\item{pattern}{search pattern}

\item{x}{your data.frame}

\item{...}{addtitional arguments for gerpl()}
}
\value{
a logical data.frame
}
\description{
Grepl applied on a data.frame
}
\examples{
matrix(letters[1:6], 2, 3) |> as.data.frame() -> a
grepl.data.frame("c", a)
grepl.data.frame("\\\\w", a)
}
