% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_box}
\alias{group_box}
\title{Plot a boxplot}
\usage{
group_box(
  tab,
  group = NULL,
  metadata = NULL,
  alpha = F,
  method = "wilcox",
  mode = 1,
  p_value = F
)
}
\arguments{
\item{tab}{your dataframe}

\item{group}{which colname choose for group or a vector}

\item{metadata}{the dataframe contains the group}

\item{alpha}{whether plot a group alphabeta}

\item{method}{test method:wilcox, tukeyHSD, LSD, (default: wilcox)}

\item{p_value}{show the p-value between each group?(when group number is 2~4)}

\item{rain}{plot a raincloud?}
}
\value{
a 'ggplot' plot object,
}
\description{
Plot a boxplot
}
\examples{
a=data.frame(a=1:18,b=runif(18,0,5))
group_box(a,group = rep(c("a","b","c"),each=6))
group_box(a[,1,drop=F],group = rep(c("a","b","c"),each=6),alpha=T,mode=3)

}
