% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{gsub.data.frame}
\alias{gsub.data.frame}
\title{Gsub applied on a data.frame}
\usage{
gsub.data.frame(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{search pattern}

\item{replacement}{a replacement for matched pattern}

\item{x}{your data.frame}

\item{...}{addtitional arguments for gerpl()}
}
\value{
a logical data.frame
}
\description{
Gsub applied on a data.frame
}
\examples{
matrix(letters[1:6], 2, 3) |> as.data.frame() -> a
gsub.data.frame("c", "a", a)
}
