% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{guolv}
\alias{guolv}
\title{Filter your data}
\usage{
guolv(tab, sum = 10, exist = 1)
}
\arguments{
\item{tab}{dataframe}

\item{sum}{the rowsum should bigger than sum(default:10)}

\item{exist}{the exist number bigger than exist(default:1)}
}
\value{
input object
}
\description{
Filter your data
}
\examples{
data(otutab)
guolv(otutab)
}
