% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{hebing2}
\alias{hebing2}
\title{Group your data}
\usage{
hebing2(otutab, group_df, margin = 2, act = "mean")
}
\arguments{
\item{otutab}{data.frame}

\item{group_df}{group data.frame with two columns (id and group). The same ID can be mapped to multiple groups.}

\item{margin}{1 for row and 2 for column(default: 2)}

\item{act}{do (default: mean)}
}
\value{
data.frame
}
\description{
Group your data
}
\examples{
data(otutab)
hebing2(otutab, data.frame(id = c("NS1", "NS2", "NS1", "NS3"), group = c("A", "A", "B", "B")))
}
