% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{how_to_set_options}
\alias{how_to_set_options}
\title{How to set options in a package}
\usage{
how_to_set_options(package = "My_package")
}
\arguments{
\item{package}{package name}
}
\value{
No return value
}
\description{
How to set options in a package
}
