% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{how_to_use_parallel}
\alias{how_to_use_parallel}
\title{How to use parallel}
\usage{
how_to_use_parallel(
  loop = function(i) {
     return(mean(rnorm(100)))
 }
)
}
\arguments{
\item{loop}{the main function}
}
\value{
No return value
}
\description{
How to use parallel
}
