% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{is.ggplot.color}
\alias{is.ggplot.color}
\title{Judge if a characteristic is Rcolor}
\usage{
is.ggplot.color(color)
}
\arguments{
\item{color}{characteristic}
}
\value{
TRUE or FALSE
}
\description{
Judge if a characteristic is Rcolor
}
\examples{
is.ggplot.color("red")
is.ggplot.color("notcolor")
is.ggplot.color(NA)
is.ggplot.color("#000")
}
