% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lib_ps}
\alias{lib_ps}
\title{Library packages or install}
\usage{
lib_ps(p_list, ..., all_yes = F)
}
\arguments{
\item{p_list}{a vector of packages list}
}
\description{
Library packages or install
}
\examples{
lib_ps("ggplot2","dplyr")
}
