% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{list_to_dataframe}
\alias{list_to_dataframe}
\title{Trans list (with NULL) to data.frame}
\usage{
list_to_dataframe(lst)
}
\arguments{
\item{lst}{list (with NULL)}
}
\value{
a data.frame
}
\description{
Trans list (with NULL) to data.frame
}
