% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{make_gitbook}
\alias{make_gitbook}
\title{Make a Gitbook using bookdown}
\usage{
make_gitbook(
  book_n,
  root_dir = "~/Documents/R/",
  mode = c("gitbook", "bs4")[1],
  author = "Asa12138",
  bib = "~/Documents/R/pc_blog/content/bib/My Library.bib",
  csl = "~/Documents/R/pc_blog/content/bib/science.csl"
)
}
\arguments{
\item{book_n}{project name}

\item{root_dir}{root directory}

\item{mode}{"gitbook","bs4"}

\item{author}{author}

\item{bib}{cite papers bib, from Zotero}

\item{csl}{cite papers format, default science.csl}
}
\value{
No return value
}
\description{
Make a Gitbook using bookdown
}
