% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{make_project}
\alias{make_project}
\title{Make a R-analysis project}
\usage{
make_project(pro_n, root_dir = "~/Documents/R/")
}
\arguments{
\item{pro_n}{project name}

\item{root_dir}{root directory}
}
\value{
No return value
}
\description{
Make a R-analysis project
}
