% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{match_df}
\alias{match_df}
\title{Match otutab and metadata}
\usage{
match_df(otutab, metadata)
}
\arguments{
\item{otutab}{otutab, rownames are features, colnames are samples}

\item{metadata}{metadata, rownames are samples}
}
\value{
list
}
\description{
Match otutab and metadata
}
\examples{
data(otutab)
match_df(otutab, metadata)
}
