% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{micro_sbatch}
\alias{micro_sbatch}
\title{Microbiome sbatch}
\usage{
micro_sbatch(
  work_dir = "/share/home/jianglab/pengchen/work/asthma/",
  step = "fastp",
  all_sample_num = 40,
  array = 1,
  partition = "cpu",
  cpus_per_task = 1,
  mem_per_cpu = "2G"
)
}
\arguments{
\item{work_dir}{work_dir}

\item{step}{"fastp","rm_human","megahit","prodigal","salmon-quant",...}

\item{all_sample_num}{all sample number}

\item{array}{array number}

\item{partition}{partition}

\item{cpus_per_task}{cpus_per_task}

\item{mem}{mem}
}
\description{
Microbiome sbatch
}
