% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{mmscale}
\alias{mmscale}
\title{Min_Max scale}
\usage{
mmscale(x, min_s = 0, max_s = 1, n = 1, plot = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{min_s}{scale min}

\item{max_s}{scale max}

\item{n}{linear transfer for n=1; the slope will change if n>1 or n<1}

\item{plot}{whether plot the transfer?}
}
\value{
a numeric vector
}
\description{
Min_Max scale
}
\examples{
x <- runif(10)
mmscale(x, 5, 10)
}
