% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{multireg}
\alias{multireg}
\title{Multiple regression/ variance decomposition analysis}
\usage{
multireg(formula, data, TopN = 3)
}
\arguments{
\item{formula}{formula}

\item{data}{dataframe}

\item{TopN}{give top variable importance}
}
\value{
ggplot
}
\description{
Multiple regression/ variance decomposition analysis
}
\examples{
data(otutab)
multireg(env1 ~ Group * ., data = metadata[, 2:7])
}
