% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multitest}
\alias{multitest}
\title{Multi-group test}
\usage{
multitest(var, group, print = T, return = F)
}
\arguments{
\item{var}{numeric vector}

\item{group}{more than two-levels group vector}

\item{return}{return which method result (tukeyHSD or LSD or wilcox?)}
}
\description{
Multi-group test
}
\examples{
multitest(c(runif(20),runif(10,2,3)),rep(c("a","b","c"),each=10))
multitest(runif(30),rep(c("a","b","c"),each=10),print=F,return="wilcox")->aa
}
