% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{my_cat}
\alias{my_cat}
\title{Show my little cat named Guo Dong which drawn by my girlfriend.}
\usage{
my_cat(mode = 1)
}
\arguments{
\item{mode}{1~2}
}
\value{
a ggplot
}
\description{
Show my little cat named Guo Dong which drawn by my girlfriend.
}
