% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_cicro}
\alias{my_cicro}
\title{My cicro plot}
\usage{
my_cicro(
  df,
  reorder = T,
  colors = NULL,
  mode = c("circlize", "chorddiag"),
  ...
)
}
\arguments{
\item{df}{dataframe}

\item{...}{\code{\link[circlize]{chordDiagram}}}
}
\value{
chordDiagram
}
\description{
My cicro plot
}
\examples{
data.frame(a=c("a","a","b","b","c"),b=c("a",LETTERS[2:5]),c=1:5)\%>\%my_cicro()
}
