% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_sankey}
\alias{my_sankey}
\title{My Sankey plot}
\usage{
my_sankey(test, ...)
}
\arguments{
\item{test}{a dataframe with hierarchical structure}

\item{...}{look for parameters in \code{\link[sankeyD3]{sankeyNetwork}}}
}
\description{
My Sankey plot
}
\examples{
data.frame(aa=rep("a",5),a=c("a","a","b","b","c"),b=c("a",LETTERS[2:5]),c=1:5)\%>\%my_sankey()
data("otutab",package = "pctax")
cbind(taxonomy,num=rowSums(otutab))[1:10,]->test
my_sankey(test)
}
