% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{my_sunburst}
\alias{my_sunburst}
\title{My Sunburst plot}
\usage{
my_sunburst(test, ...)
}
\arguments{
\item{test}{a dataframe with hierarchical structure}

\item{...}{look for parameters in \code{\link[plotly]{plot_ly}}}
}
\value{
htmlwidget
}
\description{
My Sunburst plot
}
\examples{
\donttest{
data(otutab)
cbind(taxonomy, num = rowSums(otutab))[1:10, ] -> test
my_sunburst(test)
}
}
