% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{my_voronoi_treemap}
\alias{my_voronoi_treemap}
\title{My Voronoi treemap plot}
\usage{
my_voronoi_treemap(test, ...)
}
\arguments{
\item{test}{a three-columns dataframe with hierarchical structure}

\item{...}{look for parameters in \code{\link[voronoiTreemap]{vt_d3}}}
}
\value{
htmlwidget
}
\description{
My Voronoi treemap plot
}
\examples{
\donttest{
data(otutab)
cbind(taxonomy, num = rowSums(otutab))[1:10, c(4, 7, 8)] -> test
if (requireNamespace("voronoiTreemap")) {
  my_voronoi_treemap(test)
}
}
}
