% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otutab.R
\docType{data}
\name{otutab}
\alias{otutab}
\title{test data for pcutils package}
\format{
contains an otutab, metadata and a taxonomy table.
\describe{
\item{otutab}{contians otutable rawdata}
\item{metadata}{contians metadata}
\item{taxonomy}{contians taxonomy table}
}
}
\description{
an otutab, metadata and a taxonomy table.
}
