% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcutils-package.R
\docType{package}
\name{pcutils-package}
\alias{pcutils}
\alias{pcutils-package}
\title{pcutils: Some Useful Functions for Statistics and Visualization}
\description{
It contains many useful tools or functions for statistics or visualization. Some other specialized analysis packages based on these basic functions.
}
\author{
\strong{Maintainer}: Peng Chen \email{pengchen2001@zju.edu.cn}

}
\keyword{internal}
