% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotgif}
\alias{plotgif}
\title{Plot a gif}
\usage{
plotgif(plist, file, speed = 1, ...)
}
\arguments{
\item{plist}{plot list}

\item{file}{prefix of your .gif file}

\item{speed}{1}

\item{...}{add}
}
\value{
No return value
}
\description{
Plot a gif
}
