% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{pre_number_str}
\alias{pre_number_str}
\title{Prepare a numeric string}
\usage{
pre_number_str(str, split_str = ",", continuous_str = "-")
}
\arguments{
\item{str}{a string contain ',' and '-'}

\item{split_str}{split_str ","}

\item{continuous_str}{continuous_str "-"}
}
\value{
vector
}
\description{
Prepare a numeric string
}
\examples{
pre_number_str("a1,a3,a5,a6-a10")
}
