% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{prepare_package}
\alias{prepare_package}
\title{Prepare a package}
\usage{
prepare_package(pkg_dir = ".", exclude = "print.R")
}
\arguments{
\item{pkg_dir}{defalut: "."}

\item{exclude}{vector for excluding .R files}
}
\value{
No value
}
\description{
Prepare a package
}
