% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{read.file}
\alias{read.file}
\title{Read some special format file}
\usage{
read.file(file, format = NULL, just_print = FALSE)
}
\arguments{
\item{file}{file path}

\item{format}{"blast","diamond"}

\item{just_print}{just print the file}
}
\value{
data.frame
}
\description{
Read some special format file
}
