% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{reinstall_my_packages}
\alias{reinstall_my_packages}
\title{Re-install my packages}
\usage{
reinstall_my_packages(pkgs = c("pcutils", "pctax", "MetaNet", "ReporterScore"))
}
\arguments{
\item{pkgs}{pkgs}
}
\value{
No return value
}
\description{
Re-install my packages
}
