% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove.outliers}
\alias{remove.outliers}
\title{Remove outliers}
\usage{
remove.outliers(x, factor = 1.5)
}
\arguments{
\item{x}{a numeric vector}

\item{factor}{default 1.5}
}
\description{
Remove outliers
}
\examples{
remove.outliers(c(1,10:15))
}
