% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sample_map}
\alias{sample_map}
\title{Plot the sampling map}
\usage{
sample_map(
  metadata,
  point_params = list(),
  mode = 1,
  legend_title = "Group",
  shp_file = NULL,
  crs = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{metadata}{metadata must contains "long", "lat", "Group"}

\item{point_params}{parameters parse to geom_point}

\item{mode}{1~3}

\item{legend_title}{legend title}

\item{shp_file}{a geojson file parse to \code{sf::read_sf}}

\item{crs}{crs coordinate}

\item{xlim}{xlim}

\item{ylim}{ylim}
}
\value{
map
}
\description{
Plot the sampling map
}
\examples{
sample_map(test,mode = 3,xlim = c(100,150),ylim = c(25,50))
}
