% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanxian}
\alias{sanxian}
\title{Three-line table}
\usage{
sanxian(aa, digits = 3, nrow = 10, ncol = 10, ...)
}
\arguments{
\item{aa}{a dataframe}

\item{digits}{how many digits should remain}

\item{nrow}{show how many rows}

\item{ncol}{show how many columns}

\item{...}{additional arguments e.g.(rowname=NULL)}
}
\value{
a ggplot
}
\description{
Three-line table
}
\examples{
data("otutab")
sanxian(otutab)
}
