% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scale_color_pc}
\alias{scale_color_pc}
\title{Scale a fill color}
\usage{
scale_color_pc(
  palette = NULL,
  alpha = 1,
  n = 11,
  discrete = TRUE,
  rev = FALSE,
  ...
)
}
\arguments{
\item{palette}{col1~3; or a vector of colors, you can get from: \code{RColorBrewer::brewer.pal(5,"Set2")} or \code{ggsci::pal_aaas()(5)}}

\item{alpha}{alpha}

\item{n}{how many colors you need}

\item{discrete}{TRUE or FALSE}

\item{rev}{reverse}

\item{...}{additional}
}
\value{
scale_color
}
\description{
Scale a fill color
}
