% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{split_text}
\alias{split_text}
\title{Split text into parts, each not exceeding a specified character count}
\usage{
split_text(text, nchr_each = 200)
}
\arguments{
\item{text}{Original text}

\item{nchr_each}{Maximum character count for each part}
}
\value{
List of divided parts
}
\description{
Split text into parts, each not exceeding a specified character count
}
\examples{
\donttest{
original_text <- paste0(sample(c(letters, "\n"), 400, replace = TRUE), collapse = "")
parts <- split_text(original_text, nchr_each = 200)
lapply(parts, nchar)
}
}
