% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{squash}
\alias{squash}
\title{Squash one column in a data.frame using other columns as id.}
\usage{
squash(df, column, split = ",")
}
\arguments{
\item{df}{data.frame}

\item{column}{column name, not numeric position}

\item{split}{split string}
}
\value{
data.frame
}
\description{
Squash one column in a data.frame using other columns as id.
}
\examples{
df <- data.frame(a = c(1:2, 1:2), b = letters[1:4])
squash(df, "b", ",")
}
