% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strsplit2}
\alias{strsplit2}
\title{Split Composite Names}
\usage{
strsplit2(x, split, ...)
}
\arguments{
\item{x}{character vector}

\item{split}{character to split each element of vector on, see \code{\link[base]{strsplit}}}

\item{...}{other arguments are passed to strsplit}
}
\value{
dataframe
}
\description{
Split Composite Names
}
\examples{
strsplit2(c("a;b","c;d"),";")
}
