% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tax_pie}
\alias{tax_pie}
\title{Pie plot}
\usage{
tax_pie(otutab, topN = 6, ...)
}
\arguments{
\item{otutab}{otutab}

\item{topN}{topN}

\item{...}{add}
}
\value{
ggplot
}
\description{
Pie plot
}
\examples{
tax_pie(otutab,topN = 7)
}
