% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tax_radar}
\alias{tax_radar}
\title{Radar plot}
\usage{
tax_radar(otu_time, ...)
}
\arguments{
\item{otu_time}{otutab}

\item{...}{add}
}
\description{
Radar plot
}
\examples{
\dontrun{
#data(otutab)
#tax_radar(otutab[1:20,1:3])
}
}
