% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tax_wordcloud}
\alias{tax_wordcloud}
\title{Word cloud plot}
\usage{
tax_wordcloud(str_vector)
}
\arguments{
\item{str_vector}{string vector}
}
\description{
Word cloud plot
}
\examples{
data(otutab)
tax_wordcloud(taxonomy$Genus)
}
