% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{tidai}
\alias{tidai}
\title{Replace a vector by named vector}
\usage{
tidai(x, y, fac = FALSE, keep_origin = FALSE)
}
\arguments{
\item{x}{a vector need to be replaced}

\item{y}{named vector}

\item{fac}{consider the factor?}

\item{keep_origin}{keep_origin?}
}
\value{
vector
}
\description{
Replace a vector by named vector
}
\examples{
tidai(c("a", "a", "b", "d"), c("a" = "red", b = "blue"))
tidai(c("a", "a", "b", "c"), c("red", "blue"))
tidai(c("A" = "a", "B" = "b"), c("a" = "red", b = "blue"))
tidai(factor(c("A" = "a", "B" = "b", "C" = "c")), c("a" = "red", b = "blue", c = "green"))
}
