% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{trans}
\alias{trans}
\title{Transfer your data}
\usage{
trans(df, method = "normalize", margin = 2, ...)
}
\arguments{
\item{df}{dataframe}

\item{method}{"cpm","minmax","acpm","total","log", "max", "frequency", "normalize", "range", "rank", "rrank",
"standardize", "pa", "chi.square", "hellinger", "log", "clr", "rclr", "alr"}

\item{margin}{1 for row and 2 for column(default: 2)}

\item{...}{additional}
}
\value{
data.frame
}
\description{
Transfer your data
}
\examples{

data(otutab)
trans(otutab, method = "cpm")

}
\seealso{
\code{\link[vegan]{decostand}}
}
