% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{trans_format}
\alias{trans_format}
\title{Transfer the format of file}
\usage{
trans_format(
  file,
  to_format,
  format = NULL,
  ...,
  browser = "/Applications/Microsoft Edge.app/Contents/MacOS/Microsoft Edge"
)
}
\arguments{
\item{file}{input file}

\item{to_format}{transfer to}

\item{format}{input file format}

\item{...}{additional argument}

\item{browser}{the path of Google Chrome, Microsoft Edge or Chromium in your computer.}
}
\value{
file at work directory
}
\description{
Transfer the format of file
}
