% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{translator}
\alias{translator}
\title{Translator}
\usage{
translator(words, from = "en", to = "zh", split = TRUE, verbose = TRUE)
}
\arguments{
\item{words}{words}

\item{from}{source language, default "en"}

\item{to}{target language, default "zh"}

\item{split}{split to blocks when your words are too much}

\item{verbose}{verbose}
}
\value{
vector
}
\description{
language: en, zh, jp, fra, th..., see \code{https://www.cnblogs.com/pieguan/p/10338255.html}
}
\examples{
\dontrun{
translator(c("love", "if"), from = "en", to = "zh")
}
}
