% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{twotest}
\alias{twotest}
\title{Two-group test}
\usage{
twotest(var, group)
}
\arguments{
\item{var}{numeric vector}

\item{group}{two-levels group vector}
}
\value{
No return value
}
\description{
Two-group test
}
\examples{
twotest(runif(20), rep(c("a", "b"), each = 10))
}
