% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{update_NEWS_md}
\alias{update_NEWS_md}
\title{Update the NEW.md for a package}
\usage{
update_NEWS_md(
  package_dir = ".",
  new_features = character(),
  bug_fixes = character(),
  other_changes = character(),
  ...
)
}
\arguments{
\item{package_dir}{default: "."}

\item{new_features}{new_features}

\item{bug_fixes}{bug_fixes}

\item{other_changes}{other_changes}

\item{...}{additional info}
}
\value{
No value
}
\description{
Update the NEW.md for a package
}
