% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{update_param}
\alias{update_param}
\title{Update the parameters}
\usage{
update_param(default, update)
}
\arguments{
\item{default}{default (data.frame, list, vector)}

\item{update}{update (data.frame, list, vector)}
}
\value{
same class of your input (data.frame, list or vector)
}
\description{
Keep the different parameters while use the same name in update first.
}
\examples{
update_param(list(a = 1, b = 2), list(b = 5, c = 5))

}
