% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{venn}
\alias{venn}
\alias{venn.list}
\alias{venn.data.frame}
\title{plot a general venn (upset, flower)}
\usage{
venn(object, ...)

\method{venn}{list}(aa, mode = "venn", ...)

\method{venn}{data.frame}(otutab, mode = "venn")
}
\arguments{
\item{object}{list/data.frame/pc_otu}

\item{...}{additional}

\item{mode}{"venn","venn2","upset","flower"}
}
\value{
a plot
}
\description{
plot a general venn (upset, flower)
}
\examples{
aa=list(a=1:3,b=3:7,c=2:4)
venn(aa,mode="venn")
venn(aa,mode="venn2",type="ChowRuskey")
venn(aa,mode="upset")
venn(otutab,mode="flower")
}
