% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{write_fasta}
\alias{write_fasta}
\title{Write a dataframe to fasta}
\usage{
write_fasta(df, file_path)
}
\arguments{
\item{df}{dataframe}

\item{file_path}{output file path}
}
\description{
Write a dataframe to fasta
}
