% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_theme}
\alias{add_theme}
\title{Add a global gg_theme and colors for plots}
\usage{
add_theme(set_theme = NULL)
}
\arguments{
\item{set_theme}{your theme}
}
\value{
No return value
}
\description{
Add a global gg_theme and colors for plots
}
\examples{
add_theme()
}
