% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_plot}
\alias{cor_plot}
\title{Plot correlation}
\usage{
cor_plot(
  env,
  env2 = NULL,
  mode = 1,
  method = "pearson",
  heat = T,
  addrect = 5,
  ...
)
}
\arguments{
\item{env}{dataframe1}

\item{env2}{dataframe2 (default:NULL)}

\item{mode}{plot mode (1~3)}

\item{method}{one of "pearson","kendall","spearman"}

\item{heat}{plot heatmap when columns >30}

\item{addrect}{add rect when > addrect, use for mode3}

\item{...}{}
}
\description{
Plot correlation
}
\examples{
data(otutab)
cor_plot(metadata[,3:10])
cor_plot(metadata[,3:10],mode=2)
cor_plot(t(otutab)[,1:50],mode=3)
}
