% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_plot}
\alias{cor_plot}
\title{Plot correlation}
\usage{
cor_plot(env, env2 = NULL, mode = 1, method = "pearson", heat = T, ...)
}
\arguments{
\item{env}{dataframe1}

\item{env2}{dataframe2 (default:NULL)}

\item{mode}{plot mode (1~3)}

\item{method}{one of "pearson","kendall","spearman"}
}
\description{
Plot correlation
}
\examples{
data(otutab)
cor_plot(metadata[,12:19])
cor_plot(t(otutab)[,1:50],mode=3)
}
