% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{cor_plot}
\alias{cor_plot}
\title{Plot correlation}
\usage{
cor_plot(
  env,
  env2 = NULL,
  mode = 1,
  method = "pearson",
  heat = TRUE,
  mode3_param = NULL,
  ...
)
}
\arguments{
\item{env}{dataframe1}

\item{env2}{dataframe2 (default:NULL)}

\item{mode}{plot mode (1~3)}

\item{method}{one of "pearson","kendall","spearman"}

\item{heat}{plot heatmap when columns >30}

\item{mode3_param}{parameters parse to \code{\link[corrplot]{corrplot}}}

\item{...}{for \code{\link[pheatmap]{pheatmap}}}
}
\value{
ggplot
}
\description{
Plot correlation
}
\examples{
\donttest{
if (interactive()) {
  data(otutab)
  cor_plot(metadata[, 3:10])
  cor_plot(metadata[, 3:10], mode = 2)
  cor_plot(t(otutab)[, 1:50], mode = 3, heat = FALSE)
}
}
}
