% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{count2}
\alias{count2}
\title{Like \code{uniq -c} in shell to count a vector}
\usage{
count2(df)
}
\arguments{
\item{df}{two columns: first is type, second is number}
}
\value{
two columns: first is type, second is number
}
\description{
Like \code{uniq -c} in shell to count a vector
}
\examples{
count2(data.frame(group = c("A", "A", "B", "C", "C", "A"), value = c(2, 2, 2, 1, 3, 1)))
}
