% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{explode}
\alias{explode}
\title{Explode a data.frame if there are split charter in one column}
\usage{
explode(df, column, split = ",")
}
\arguments{
\item{df}{data.frame}

\item{column}{column}

\item{split}{split string}
}
\value{
data.frame
}
\description{
Explode a data.frame if there are split charter in one column
}
\examples{
\donttest{
df <- data.frame(a = 1:2, b = c("a,b", "c"), c = 3:4)
explode(df, "b", ",")
}
}
