% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_cols}
\alias{get_cols}
\title{Get n colors}
\usage{
get_cols(n = 11, pal = "col1")
}
\arguments{
\item{n}{how many colors you need}

\item{pal}{col1~3; or a vector of colors, you can get from: \code{RColorBrewer::brewer.pal(5,"Set2")} or \code{ggsci::pal_aaas()(5)}}
}
\value{
a vector of n colors
}
\description{
Get n colors
}
\examples{
get_cols(10, "col2") -> my_cols
scales::show_col(my_cols)
\donttest{
scales::show_col(get_cols(15, RColorBrewer::brewer.pal(5, "Set2")))
scales::show_col(get_cols(15, ggsci::pal_aaas()(5)))
}
}
