% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gghist}
\alias{gghist}
\title{gg histogram}
\usage{
gghist(x, text_pos = c(0.8, 0.8), ...)
}
\arguments{
\item{x}{vector}

\item{text_pos}{text postion, default is c(0.8, 0.8)}

\item{...}{parameters parse to \code{\link[ggpubr]{gghistogram}}}
}
\value{
ggplot
}
\description{
gg histogram
}
\examples{
if (requireNamespace("ggpubr")) {
  gghist(rnorm(100))
}
}
