% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gghuan}
\alias{gghuan}
\title{Plot a doughnut chart}
\usage{
gghuan(
  tab,
  reorder = TRUE,
  mode = "1",
  topN = 5,
  name = TRUE,
  percentage = TRUE
)
}
\arguments{
\item{tab}{two columns: first is type, second is number}

\item{reorder}{reorder by number?}

\item{mode}{plot style, 1~3}

\item{topN}{plot how many top items}

\item{name}{label the name}

\item{percentage}{label the percentage}
}
\value{
a ggplot
}
\description{
Plot a doughnut chart
}
\examples{
a <- data.frame(type = letters[1:6], num = c(1, 3, 3, 4, 5, 10))
gghuan(a) + ggplot2::scale_fill_manual(values = get_cols(6, "col3"))
b <- data.frame(type = letters[1:12], num = c(1, 3, 3, 4, 15, 10, 35, 2:6))
gghuan(b) + ggplot2::theme(legend.position = "right")
}
