% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gghuan2}
\alias{gghuan2}
\title{gghuan2 for multi-doughnut chart}
\usage{
gghuan2(
  tab = NULL,
  space_width = 0.2,
  name = TRUE,
  percentage = FALSE,
  text_params = NULL,
  bar_params = NULL
)
}
\arguments{
\item{tab}{a dataframe with hierarchical structure}

\item{space_width}{the space width between doughnuts (0~1).}

\item{name}{label the name}

\item{percentage}{label the percentage}

\item{text_params}{parameters parse to \code{\link[ggplot2]{geom_text}}}

\item{bar_params}{parameters parse to \code{\link[ggplot2]{geom_rect}}}
}
\value{
a ggplot
}
\description{
gghuan2 for multi-doughnut chart
}
\examples{
data.frame(
    a = c("a", "a", "b", "b", "c"), b = c("a", LETTERS[2:5]), c = rep("a", 5),
    number = 1:5
) \%>\% gghuan2()
}
