% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_box}
\alias{group_box}
\title{Plot a boxplot}
\usage{
group_box(
  tab,
  col = NULL,
  metadata = NULL,
  group = NULL,
  alpha = T,
  method = "tukeyHSD",
  rain = F,
  p_value = F
)
}
\arguments{
\item{tab}{your dataframe}

\item{col}{which column choose for value}

\item{metadata}{the dataframe contains the group}

\item{group}{which colname choose for group or a vector}

\item{alpha}{whether plot a group alphabeta}

\item{method}{test method:tukeyHSD,LSD, (default: tukeyHSD)}

\item{rain}{plot a raincloud?}

\item{p_value}{show the p-value between each group?(when group number is 2~4)}
}
\value{
a 'ggplot' plot object,
}
\description{
Plot a boxplot
}
\examples{
a=data.frame(a=1:18,b=runif(18,0,5))
group_box(a,col = 2,group = rep(c("a","b","c"),each=6))
group_box(a,col = 1,group = rep(c("a","b","c"),each=6),alpha=F,rain=T,each=T)

}
