% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{group_box}
\alias{group_box}
\title{Plot a boxplot}
\usage{
group_box(
  tab,
  group = NULL,
  metadata = NULL,
  mode = 1,
  group_order = NULL,
  facet_order = NULL,
  alpha = F,
  method = "wilcox",
  alpha_param = list(color = "red"),
  p_value1 = F,
  p_value2 = F,
  stat_compare_means_param = NULL,
  trend_line = F,
  trend_line_param = list(color = "blue")
)
}
\arguments{
\item{tab}{your dataframe}

\item{group}{which colname choose for group or a vector}

\item{metadata}{the dataframe contains the group}

\item{mode}{1~3, plot style}

\item{group_order}{the order of x group}

\item{facet_order}{the order of the facet}

\item{alpha}{whether plot a group alphabeta by test of method}

\item{method}{test method:wilcox, tukeyHSD, LSD, (default: wilcox), see \code{\link{multitest}}}

\item{alpha_param}{parameters parse to \code{\link[ggplot2]{geom_text}}}

\item{p_value1}{multi-test of all group}

\item{p_value2}{two-test of each pair}

\item{stat_compare_means_param}{parameters parse to \code{\link[ggpubr]{stat_compare_means}}}

\item{trend_line}{add a trend line}

\item{trend_line_param}{parameters parse to \code{\link[ggplot2]{geom_smooth}}}
}
\value{
a 'ggplot' plot object,
}
\description{
Plot a boxplot
}
\examples{
a <- data.frame(a = 1:18, b = runif(18, 0, 5))
group_box(a, group = rep(c("a", "b", "c"), each = 6), p_value1 = FALSE, p_value2 = TRUE)
multitest(a$a, group = rep(c("a", "b", "c"), each = 6))
group_box(a[, 1, drop = F], group = rep(c("a", "b", "c"), each = 6), p_value2 = T, mode = 3,stat_compare_means_param=list(comparisons = list(c("a", "b")))) + mytheme
}
