% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_box}
\alias{group_box}
\title{Plot a boxplot}
\usage{
group_box(
  tab,
  group = NULL,
  metadata = NULL,
  mode = 1,
  facet = T,
  alpha = F,
  method = "wilcox",
  p_value1 = F,
  p_value2 = F,
  comparisons = NULL,
  ...
)
}
\arguments{
\item{tab}{your dataframe}

\item{group}{which colname choose for group or a vector}

\item{metadata}{the dataframe contains the group}

\item{mode}{1~3}

\item{facet}{facet when only one var?}

\item{alpha}{whether plot a group alphabeta}

\item{method}{test method:wilcox, tukeyHSD, LSD, (default: wilcox)}

\item{p_value1}{multi-test of all group}

\item{p_value2}{two-test of each pair}

\item{comparisons}{comparison pairs, as list, e.g. list(c("a","b"))}
}
\value{
a 'ggplot' plot object,
}
\description{
Plot a boxplot
}
\examples{
a=data.frame(a=1:18,b=runif(18,0,5))
group_box(a,group = rep(c("a","b","c"),each=6),p_value1=F,p_value2=T)
multitest(a$a,group = rep(c("a","b","c"),each=6))
group_box(a[,1,drop=F],group = rep(c("a","b","c"),each=6),p_value2=T,mode=3,comparisons=list(c("a","b")))

}
