% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{lib_ps}
\alias{lib_ps}
\title{Attach packages or install packages have not benn installed}
\usage{
lib_ps(p_list, ..., all_yes = FALSE, library = TRUE)
}
\arguments{
\item{p_list}{a vector of packages list}

\item{...}{packages}

\item{all_yes}{all install try set to yes?}

\item{library}{should library the package or just get Namespace ?}
}
\value{
No return value
}
\description{
Attach packages or install packages have not benn installed
}
