% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{multitest}
\alias{multitest}
\title{Multi-groups test}
\usage{
multitest(var, group, print = T, return = F)
}
\arguments{
\item{var}{numeric vector}

\item{group}{more than two-levels group vector}

\item{print}{whether print the result}

\item{return}{return which method result (tukeyHSD or LSD or wilcox?)}
}
\description{
anova (parametric) and kruskal.test (non-parametric). Perform one-way ANOVA test comparing multiple groups.
LSD and TukeyHSD are post hoc test of anova.
dunn and nemenyi are post hoc test of kruskal.test.
ttest or wilcox is just perform wilcox-test in each two group (no p.adjust).
}
\examples{
\dontrun{
#multitest(c(runif(20), runif(10, 2, 3)), rep(c("a", "b", "c"), each = 10))
#multitest(runif(30), rep(c("a", "b", "c"), each = 10), print = F, return = "wilcox") -> aa
}
}
