% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{multitest}
\alias{multitest}
\title{Multi-groups test}
\usage{
multitest(var, group, print = TRUE, return = FALSE)
}
\arguments{
\item{var}{numeric vector}

\item{group}{more than two-levels group vector}

\item{print}{whether print the result}

\item{return}{return which method result (tukeyHSD or LSD or wilcox?)}
}
\value{
No value or a dataframe.
}
\description{
anova (parametric) and kruskal.test (non-parametric). Perform one-way ANOVA test comparing multiple groups.
LSD and TukeyHSD are post hoc test of anova.
dunn and nemenyi are post hoc test of kruskal.test.
t.test or wilcox is just perform t.test or wilcox.test in each two group (no p.adjust).
}
\examples{
multitest(runif(30), rep(c("A", "B", "C"), each = 10), print = FALSE, return = "wilcox") -> aa
}
