% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_lm}
\alias{my_lm}
\title{Fit a linear model and plot}
\usage{
my_lm(tab, var, metadata = NULL, facet = T, ...)
}
\arguments{
\item{tab}{your dataframe}

\item{var}{which colname choose for var or a vector}

\item{metadata}{the dataframe contains the var}

\item{facet}{facet when only one var?}

\item{...}{}
}
\value{
a ggplot
}
\description{
Fit a linear model and plot
}
\examples{
my_lm(runif(50),var=1:50)
my_lm(c(1:50)+runif(50,0,5),var=1:50)
}
