% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_lm}
\alias{my_lm}
\title{Fit a lm and plot}
\usage{
my_lm(y, x = NULL, ...)
}
\arguments{
\item{y}{a numeric vector or a two-columns dataframe}

\item{x}{a numeric vector or NULL}
}
\value{
a ggplot
}
\description{
Fit a lm and plot
}
\examples{
my_lm(runif(50),1:50)
my_lm(c(1:50)+runif(50,0,5),1:50)
}
