% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{my_sankey}
\alias{my_sankey}
\title{My Sankey plot}
\usage{
my_sankey(test, mode = c("sankeyD3", "ggsankey"), space = 1, ...)
}
\arguments{
\item{test}{a dataframe with hierarchical structure}

\item{mode}{"sankeyD3","ggsankey"}

\item{space}{space width for ggsankey}

\item{...}{look for parameters in \code{\link[sankeyD3]{sankeyNetwork}}}
}
\value{
ggplot or htmlwidget
}
\description{
My Sankey plot
}
\examples{
\donttest{
data.frame(a=c("a","a","b","b","c"),aa=rep("a",5),b=c("a",LETTERS[2:5]),c=1:5)\%>\%
   my_sankey(.,"gg",num=TRUE)
data(otutab)
cbind(taxonomy,num=rowSums(otutab))[1:10,]->test
my_sankey(test)->p
}
}
