% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{my_treemap}
\alias{my_treemap}
\title{My Treemap plot}
\usage{
my_treemap(test, ...)
}
\arguments{
\item{test}{a three-columns dataframe with hierarchical structure}

\item{...}{look for parameters in \code{\link[plotly]{plot_ly}}}
}
\value{
htmlwidget
}
\description{
My Treemap plot
}
\examples{
\donttest{
data(otutab)
cbind(taxonomy, num = rowSums(otutab))[1:10, c(4, 7, 8)] -> test
my_treemap(test)
}
}
