% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{plot.coefficients}
\alias{plot.coefficients}
\title{Plot coefficients as a bar chart or lollipop chart}
\usage{
\method{plot}{coefficients}(x, mode = 1, number = FALSE, x_order = NULL, ...)
}
\arguments{
\item{x}{The coefficients to be plotted.}

\item{mode}{The mode of the plot: 1 for bar chart, 2 for lollipop chart.}

\item{number}{show number}

\item{x_order}{order of variables}

\item{...}{add}
}
\value{
ggplot
}
\description{
This function takes the coefficients and generates a plot to visualize their magnitudes.
}
