% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{plotgif}
\alias{plotgif}
\title{Plot a gif}
\usage{
plotgif(plist, file, mode = "gif")
}
\arguments{
\item{plist}{plot list}

\item{file}{prefix of your .gif file}

\item{mode}{"gif" or "html"}
}
\value{
No return value
}
\description{
Plot a gif
}
