% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotpdf}
\alias{plotpdf}
\title{Plot a multi-pages pdf}
\usage{
plotpdf(plist, file = "new", width = 8, height = 7, ...)
}
\arguments{
\item{plist}{plot list}

\item{file}{prefix of your .pdf file}

\item{width}{width}

\item{height}{height}

\item{...}{additional arguments}
}
\description{
Plot a multi-pages pdf
}
