% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{read.file}
\alias{read.file}
\title{Read some special format file}
\usage{
read.file(
  file,
  format = NULL,
  just_print = FALSE,
  all_yes = FALSE,
  density = 120,
  ...
)
}
\arguments{
\item{file}{file path}

\item{format}{"blast", "diamond", "fa", "fasta", "fna", "faa", "bib", "gff", "gtf","jpg", "png", "pdf", "svg"...}

\item{just_print}{just print the file}

\item{all_yes}{all_yes?}

\item{density}{the resolution for reading pdf or svg}

\item{...}{additional arguments}
}
\value{
data.frame
}
\description{
Read some special format file
}
