% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{rgb2code}
\alias{rgb2code}
\title{Transform a rgb vector to a Rcolor code}
\usage{
rgb2code(x, rev = FALSE)
}
\arguments{
\item{x}{vector or three columns data.frame}

\item{rev}{reverse,transform a Rcolor code to a rgb vector}
}
\value{
Rcolor code like "#69C404"
}
\description{
Transform a rgb vector to a Rcolor code
}
\examples{
rgb2code(c(12, 23, 34))
rgb2code("#69C404", rev = TRUE)
}
