% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{sanxian}
\alias{sanxian}
\title{Three-line table}
\usage{
sanxian(
  df,
  digits = 3,
  nrow = 10,
  ncol = 10,
  fig = FALSE,
  mode = 1,
  background = "#D7261E",
  ...
)
}
\arguments{
\item{df}{a data.frame}

\item{digits}{how many digits should remain}

\item{nrow}{show how many rows}

\item{ncol}{show how many columns}

\item{fig}{output as a figure}

\item{mode}{1~2}

\item{background}{background color}

\item{...}{additional arguments e.g.(rows=NULL)}
}
\value{
a ggplot
}
\description{
Three-line table
}
\examples{
\donttest{
data(otutab)
sanxian(otutab)
}
}
