% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{strsplit2}
\alias{strsplit2}
\title{Split Composite Names}
\usage{
strsplit2(x, split, colnames = NULL, ...)
}
\arguments{
\item{x}{character vector}

\item{split}{character to split each element of vector on, see \code{\link[base]{strsplit}}}

\item{colnames}{colnames for the result}

\item{...}{other arguments are passed to \code{\link[base]{strsplit}}}
}
\value{
data.frame
}
\description{
Split Composite Names
}
\examples{
strsplit2(c("a;b", "c;d"), ";", colnames = c("col1", "col2"))
}
