% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{tax_radar}
\alias{tax_radar}
\title{Radar plot}
\usage{
tax_radar(group_df, ...)
}
\arguments{
\item{group_df}{group_df}

\item{...}{add}
}
\value{
ggplot
}
\description{
Radar plot
}
\examples{
\donttest{
if (interactive()) {
  data(otutab)
  tax_radar(otutab[1:6, 1:4])
}
}
}
