% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{tidai}
\alias{tidai}
\title{Replace a vector by named vector}
\usage{
tidai(x, y, fac = FALSE)
}
\arguments{
\item{x}{a vector need to be replaced}

\item{y}{named vector}

\item{fac}{consider the factor?}
}
\value{
vector
}
\description{
Replace a vector by named vector
}
\examples{
tidai(c("a", "a", "b"), c("a" = "red", b = "blue"))
tidai(c("a", "a", "b", "c"), c("red", "blue"))
}
