% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{toXY}
\alias{toXY}
\title{Transfer Geographical latitude and longitude to XY(m)}
\usage{
toXY(geo)
}
\arguments{
\item{geo}{a two-columns dataframe, first is latitude, second is longitude}
}
\value{
data.frame
}
\description{
Transfer Geographical latitude and longitude to XY(m)
}
\examples{
\donttest{
if (interactive()) {
  data.frame(row.names = letters[1:18], x = runif(18, 30, 35), y = runif(18, 40, 45)) -> geo
  toXY(geo)
}
}
}
