% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{translator}
\alias{translator}
\title{translator}
\usage{
translator(words, mode = "e2z")
}
\arguments{
\item{words}{words}

\item{mode}{"e2z","z2e"}
}
\value{
vector
}
\description{
translator
}
\examples{
\dontrun{
translator(c("love", "if"), mode = "e2z")
}
}
