% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{triangp}
\alias{triangp}
\title{Triangle plot}
\usage{
triangp(group_df, class = NULL)
}
\arguments{
\item{group_df}{group_df}

\item{class}{point color}
}
\description{
Triangle plot
}
\examples{
\donttest{
data(otutab)
hebing(otutab,metadata$Group,act = 'mean')->tmp
triangp(tmp,class=taxonomy$Phylum)
}
}
