% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{triangp}
\alias{triangp}
\title{Triangle plot}
\usage{
triangp(otutab, group, scale = F, class = NULL)
}
\arguments{
\item{otutab}{otutab}

\item{group}{group}

\item{scale}{default:F}

\item{class}{point color}
}
\description{
Triangle plot
}
\examples{
data(otutab)
triangp(otutab,metadata$Group,class=taxonomy$Phylum,scale=TRUE)
}
