% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{venn}
\alias{venn}
\alias{venn.list}
\alias{venn.data.frame}
\title{Plot a general venn (upset, flower)}
\usage{
venn(...)

\method{venn}{list}(aa, mode = "venn", elements_label = TRUE, ...)

\method{venn}{data.frame}(otutab, mode = "venn", elements_label = TRUE, ...)
}
\arguments{
\item{...}{add}

\item{aa}{list}

\item{mode}{"venn","venn2","upset","flower"}

\item{elements_label}{logical, show elements label in network?}

\item{otutab}{table}
}
\value{
a plot

a plot

a plot
}
\description{
Plot a general venn (upset, flower)
}
\examples{
\donttest{
aa <- list(a = 1:3, b = 3:7, c = 2:4)
venn(aa, mode = "venn")
venn(aa, mode = "network")
venn(aa, mode = "upset")
data(otutab)
venn(otutab, mode = "flower")
}
}
