% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_tools.R
\name{write_fasta}
\alias{write_fasta}
\title{Write a data.frame to fasta}
\usage{
write_fasta(df, file_path, str_per_line = 70)
}
\arguments{
\item{df}{data.frame}

\item{file_path}{output file path}

\item{str_per_line}{how many base or animo acid in one line, if NULL, one sequence in one line.}
}
\value{
No return value
}
\description{
Write a data.frame to fasta
}
